### 2024-09-02 #7167 @andreanini
Suggested reviewers:

Mike Kestemont (@ mikekestemont)
George Mikros (@ gmikros)
Simone Rebora (@ SimoneRebora)
Jean-Baptiste Camps (@ Jean-Baptiste-Camps)
Artjoms Šeļa (@ perechen)
Ben Nagy (@ bnagy)
