### 2024-09-06 #7167 @samhforbes
Hi @andreanini thanks for submitting to JOSS.
For packages of this size we will often run checks to see whether it is in scope for JOSS according to our submission requirements, so I will run a scope check now.
If you have any particular thoughts as to why it should be considered in scope you can comment that below, otherwise we will let you know when this is complete.
