### 2024-09-06 #7167 @andreanini
Hi @samhforbes, can you just clarify what you mean by "size"? The package has more than 1,000 lines or R code and it's the result of more than three months of work over 125 commits. I can only see these two criteria in your guidelines regarding size. If there are any other concerns regarding the other parameters please let me know and I'll see if I can provide any clarifications. The package has not been cited yet because it's new but it has the potential of being widely cited. The publication of the paper in JOSS would also help in that regard, of course.

I'm concerned by the fact that I cannot see any paper in the closed issues that has been flagged by query-score and then accepted, which seems to suggest that this might amount to a desk-rejection effectively.
