### 2024-09-09 #7167 @danielskatz
@andreanini - My guess is that about 10-20% of scope queried submissions get accepted for review, and about 75-80% of those do get accepted.
