### 2024-09-11 #7167 @andreanini
Thank you for your clarification, @danielskatz. Before I submitted the paper my impression was that my contribution fitted the criteria and I thought the stats about lines of code confirmed this. If you could let me know what other concerns you have I can try to provide some comments to address them.
