### 2024-09-20 #7167 @samhforbes
Hi @thomaskrause, @stefanocoretta and @mikekestemont would any of you be willing to review this submission for us? We follow these review guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
