### 2024-09-20 232cb10
fix bug in path to print
