### 2024-09-20 0d38dbd
add experimental function lambdaG_patterns
