+++ tests/testthat/test-lambdaG_patterns.R
@@ -1,0 +1,10 @@
+test_that("lambdaG patterns works", {
+
+  q.data <- quanteda::corpus_trim(enron.sample[1], "sentences", max_ntoken = 10) |>
+    quanteda::tokens("sentence")
+  k.data <- enron.sample[2:5] |> quanteda::tokens("sentence")
+  ref.data <- enron.sample[6:ndoc(enron.sample)] |> quanteda::tokens("sentence")
+  set.seed(2)
+  lambdaG_patterns(q.data, k.data, ref.data, r = 2) |> expect_snapshot()
+
+})
