+++ NAMESPACE
@@ -9,6 +9,7 @@ export(delta)
 export(density_plot)
 export(impostors)
 export(lambdaG)
+export(lambdaG_patterns)
 export(lambdaG_visualize)
 export(most_similar)
 export(ngram_tracing)
