### 2024-10-05 43453de
add leave-one-author-out and optional progress bar to performance()
