+++ NEWS.md
@@ -1,5 +1,9 @@
 # idiolect (development version)
 
+* changed progress bar in performance() to optional
+* added the option to leave-one-out by author in performance()
+* bug fixes
+
 # idiolect 1.0.1
 
 * Fixed issues after CRAN review.
