### 2024-10-09 2738a56
changing detokenize to add sentence boundaries if needed
