### 2024-10-09 1c4023a
concordance to accept sentence tokens as objects and search using sentence boundaries
