+++ man/concordance.Rd
@@ -15,13 +15,13 @@ concordance(
 )
 }
 \arguments{
-\item{q.data}{A \code{quanteda} corpus object, such as the output of \code{\link[=create_corpus]{create_corpus()}}.}
+\item{q.data}{A \code{quanteda} corpus object, such as the output of \code{\link[=create_corpus]{create_corpus()}}, or a tokens object with tokens being sentences, such as the output of \code{\link[=tokenize_sents]{tokenize_sents()}}.}
 
-\item{k.data}{A \code{quanteda} corpus object, such as the output of \code{\link[=create_corpus]{create_corpus()}}.}
+\item{k.data}{A \code{quanteda} corpus object, such as the output of \code{\link[=create_corpus]{create_corpus()}}, or a tokens object with tokens being sentences, such as the output of \code{\link[=tokenize_sents]{tokenize_sents()}}.}
 
-\item{reference.data}{A \code{quanteda} corpus object, such as the output of \code{\link[=create_corpus]{create_corpus()}}. This is optional.}
+\item{reference.data}{A \code{quanteda} corpus object, such as the output of \code{\link[=create_corpus]{create_corpus()}}, or a tokens object with tokens being sentences, such as the output of \code{\link[=tokenize_sents]{tokenize_sents()}}. This is optional.}
 
-\item{search}{A string. It can be any sequence of characters and it also accepts the use of * as a wildcard.}
+\item{search}{A string. It can be any sequence of characters and it also accepts the use of * as a wildcard. The special tokens for sentence boundaries are '\emph{BOS}' for beginning of sentence and '\emph{EOS}' for end of sentence.}
 
 \item{token.type}{Choice between "word" (default), which searches for word or punctuation mark tokens, or "character", which instead uses a single character search.}
 
@@ -44,5 +44,8 @@ concordance(enron.sample[1], enron.sample[2], enron.sample[3:49], "want * to", t
 #searching character sequences with wildcards
 concordance(enron.sample[1], enron.sample[2], enron.sample[3:49], "help*", token.type = "character")
 
+#using sentences
+enron.sents <- tokens(enron.sample, "sentence")
+concordance(enron.sents[1], enron.sents[2], enron.sents[3:49], ". _EOS_", token.type = "word")
 
 }
