### 2024-10-09 4d339a5
add possibility to visualise negative values of lambdaG in html
