+++ NEWS.md
@@ -1,5 +1,6 @@
 # idiolect (development version)
 
+* add the possibility to visualize negative lambdaG values in an html file
 * concordance() now can take sentences as input and will also show sentence boundaries
 * changed progress bar in performance() to optional
 * added the option to leave-one-out by author in performance()
