### 2024-11-18 #7167 @andreanini
Sure. If you haven't already, you could try Valerio Gherardi (@ vgherard), Maciej Eder (https://maciejeder.org), Patrick Juola (https://www.duq.edu/faculty-and-staff/patrick-juola.php), or Joanna Byszuk (https://joannaby.github.io).
