### 2024-11-25 #7167 @samhforbes
@vgherard, @cmaimone would any of you be willing to review this submission for us? We follow these review guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
