### 2024-11-27 #7167 @samhforbes
Hi @vgherard thanks for the transparency. I think I'd be tempted to agree that this is a potential COI. I really appreciate the thoughtful response though!
