### 2024-12-09 #7575 @editorialbot
[REVIEW]: idiolect: An R package for forensic authorship analysis
**Submitting author:** @andreanini (<a href="http://orcid.org/0000-0003-4218-5130">Andrea Nini</a>)
**Repository:** https://github.com/andreanini/idiolect
**Branch with paper.md** (empty if default branch): joss
**Version:** v.1.0.1
