### 2025-01-02 7672f07
Typo correction in paper
offer >> often
