+++ paper.md
@@ -38,7 +38,7 @@ Within this context, `idiolect` is an R package that contains functions to pre-p
 
 Most significantly, what sets `idiolect` apart is its use of the Likelihood Ratio Framework. Through a suite of functions, `idiolect` facilitates the calibration of likelihood ratios from the results of any of the authorship analysis functions and then the assessment of the performance of this likelihood ratio using standard performance metrics, such as the $C_{llr}$ [@ramos2013].
 
-Another novelty in `idiolect` is that the package also offers functions that aid the *post-hoc* interpretation of the results. Computational authorship analysis techniques are offer hard to interpret by the analyst. Although this is true, for example, for algorithms such as the *Impostors Method* that are based on the frequency of short sequences of characters, `idiolect` facilitates interpretation by returning the most important features and allowing the user to see these features in context. For *LambdaG*, a purpose-built function can return a colour-coded heat map of a text highlighting the words or constructions that influenced the results.
+Another novelty in `idiolect` is that the package also offers functions that aid the *post-hoc* interpretation of the results. Computational authorship analysis techniques are often hard to interpret by the analyst. Although this is true, for example, for algorithms such as the *Impostors Method* that are based on the frequency of short sequences of characters, `idiolect` facilitates interpretation by returning the most important features and allowing the user to see these features in context. For *LambdaG*, a purpose-built function can return a colour-coded heat map of a text highlighting the words or constructions that influenced the results.
 
 Although `idiolect` has been designed for research in authorship analysis, stylometry, digital humanities, and forensic linguistics, it can also be used effectively to run analyses for real-life forensic linguistics casework. The code being open source is particularly important in a forensic context to allow opposing experts to replicate the analysis and scrutinize the procedure in full.
 
