+++ paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 Authorship Analysis is defined as the task of determining the likelihood that a certain person is the author of a certain set of questioned texts. This determination is done by analysing the language of the questioned texts and the language of samples produced by the candidate author or authors. This kind of analysis is often applied in the context of literary problems (e.g. Robert Galbraith as J.K. Rowling's alias [@juola2015], the identity of Elena Ferrante [@tuzzi2018]), historical problems (e.g. Lincoln's Bixby letter [@grieve2019], the Jack the Ripper letters [@nini2018], the writings of Julius Caesar [@kestemont2016]) or forensic contexts (e.g. the *devil strip* ransom letter [@leonard2005], the Amanda Birks murder [@grant2013] or the Ayia Napa rape statements [@donlan2022]).
 
-Especially when dealing with a forensic context, then best practice is to carry out authorship analysis within the Bayesian Likelihood Ratio Framework for expressing evidence in forensic science, which is logically aligned with the role of the expert witness in a court of law. Rather than expressing a final binary judgement (same author vs. different author, or author A vs. author B), the framework instead leads the analyst to express the strength of the linguistic evidence in favour or against a set of two competing hypotheses, for example:
+Especially when dealing with a forensic context, the best practice is to carry out authorship analysis within the Bayesian Likelihood Ratio Framework for expressing evidence in forensic science, which is logically aligned with the role of the expert witness in a court of law. Rather than expressing a final binary judgement (same author vs. different author, or author A vs. author B), the framework instead leads the analyst to express the strength of the linguistic evidence in favour or against a set of two competing hypotheses, for example:
 
 | $H_p$: The candidate author and the author of the questioned text are the same individual.
 | $H_d$: The candidate author and the author of the questioned text are two different individuals.
