### 2025-01-02 #7575 @cmaimone
There is a mild performance claim in the paper ("By being based on quanteda, the functions in idiolect can efficiently handle very large matrices and can therefore process data quickly or handle very large datasets. This factor is a significant advantage compared to other R packages for authorship analysis such as stylo [@eder2016].") that isn't supported with any tests.  The speed of quanteda is, however, well documented, so I don't think this is necessarily an issue.  Noting for @samhforbes in case it matters.
