### 2025-01-02 andreanini/idiolect#6 @cmaimone
readtext dependency
If readtext is required for one of the main package functions (create_corpus), it seems like it should be a dependency and installed with your package (instead of just suggested).  Unless there's some other way to use the package that wouldn't require this?

re: https://github.com/openjournals/joss-reviews/issues/7575
