### 2025-01-02 andreanini/idiolect#7 @cmaimone
create_corpus error text
I (intentionally) tried to create a corpus using files that weren't named as specified in the documentation (just called doc1.txt, doc2.txt, etc.).  The error message is:
```
Error in names(docvar) <- `*vtmp*` :
  'names' attribute [2] must be the same length as the vector [1]
```
Which doesn't help the user solve the problem.  Suggestion to consider updating the relevant error messages (there may be more than just this one) for this

re: https://github.com/openjournals/joss-reviews/issues/7575
