### 2025-01-02 andreanini/idiolect#8 @cmaimone
Validation documention
I got lost in the validation documentation https://andreanini.github.io/idiolect/articles/idiolect.html#validation because the first code chunk had validation being K + R, but then it had Q and K as the validation subsets.  I wasn't sure what I was supposed to be doing at that step.

Related, it would help in the data labelling section https://andreanini.github.io/idiolect/articles/idiolect.html#data-labelling to remind people what's supposed to be in K, Q, and R, because I wasn't following the mix of known and unknown and the different authors -- I can see it now looking at the corpus, but I wasn't following initially.  Maybe a bit more description of what's in the example enron corpus?  I think maybe there's a conceptual issue with labeling documents by author -- that the author is listed as KW but KW is not really known as the author?  I'm not sure where the known and unknown labels came in.

re: https://github.com/openjournals/joss-reviews/issues/7575
