### 2025-01-02 andreanini/idiolect#9 @cmaimone
impostors() warning
When running the vignette, I get:

```
In FUN(X[[i]], ...) :
  The k parameter is greater than or equal to the number of available impostors.
```

I wouldn't normally expect a vignette to produce a warning message.  Do you want an example that produces a warning?

re: https://github.com/openjournals/joss-reviews/issues/7575
