### 2025-01-02 andreanini/idiolect#10 @cmaimone
documentation inconsistency
The colors in the plot and the text don't match:

re: https://github.com/openjournals/joss-reviews/issues/7575
