### 2025-01-02 andreanini/idiolect#8 @cmaimone
This may just be my lack of subject matter knowledge, but after getting through the validation section, I don't know what the goal of that step was.  Maybe add a few sentences at the beginning of that section about what needs to be validated and why?

And at the end, a little more detail on how you know that it (whatever it is) is valid would also be helpful.  It's useful at the end where you point to values of specific parameters that you're looking at, but I still wasn't following entirely.  I wasn't familiar with some of the language used to talk about LLR -- is this a log-likelihood ratio _test_?  or a _model?  A ratio itself is just a number, so I was confused by sentences like "there is enough information in the LLR" -- because how is there information in a number?  These are small language issues that may reflect norms in your field that I'm not familiar with though.
