### 2025-01-02 andreanini/idiolect#11 @cmaimone
Community/contributing guidelines
I didn't see guidelines for contributing to the package and getting help/reporting issues.

re: https://github.com/openjournals/joss-reviews/issues/7575
