### 2025-01-02 #7575 @cmaimone
@stefanocoretta - are you familiar with the analysis methods that this package provides?  I got a bit lost in the vignette, but that may just be because this isn't a type of analysis I do (I'm familiar with this type of analysis in general at a high level, but not at the level of knowing the recent papers cited, specific metrics, etc.).  I think it would be helpful for someone more familiar with the field to assess whether the vignette and documentation are sufficient for likely users (they weren't quite sufficient for me, as a non-expert user).
