### 2025-01-06 #7575 @samhforbes
Hi @cmaimone is there a particular part of it that is unfamiliar?
