### 2025-01-09 #7575 @stefanocoretta
@cmaimone I am vaguely familiar with forensic linguistics but I could follow the reasoning in the vignette. I think the package is not necessarily intended for those who have no background in that type of analysis (as in, they would have to fill that gap with other sources, rather than with the vignettes), but maybe this can be made explicit at the beginning of the vignette?
