### 2025-03-07 fdb5812
remove lambdaG_patterns
