+++ R/globals.R
@@ -1,2 +1,2 @@
 #the following line avoids the note re variables when checking
-utils::globalVariables(c("author", "token", "doc_id", "POSnoise", "Var1", "Var2", "llr", "target", "prosecution_prior_probs", "LR", "prior_odds", "post_odds", "score", "pattern", "sentence_id", "token_id", "sentence_llr", "desc", "sentence_lambdaG", "original_docid", "token_contribution", "sentence_contribution", "is_emoji", "keyword", "chull", "binomial"))
+utils::globalVariables(c("author", "token", "doc_id", "POSnoise", "Var1", "Var2", "llr", "target", "prosecution_prior_probs", "LR", "prior_odds", "post_odds", "score", "pattern", "sentence_id", "token_id", "sentence_llr", "desc", "sentence_lambdaG", "original_docid", "token_contribution", "sentence_contribution", "is_emoji", "keyword", "chull", "binomial", "color", "K"))
