+++ dev/pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 3.1.11
-pkgdown: 2.1.0
+pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   idiolect: idiolect.html
-last_built: 2024-08-29T16:24Z
+last_built: 2025-03-14T14:43Z
 urls:
   reference: https://andreanini.github.io/idiolect/reference
   article: https://andreanini.github.io/idiolect/articles
