### 2025-04-11 3095e97
fixed bug in ngram_tracing() when performing tests with multiple known authors
