+++ tests/testthat/_snaps/ngram_tracing.md
@@ -5,26 +5,26 @@
     Output
                                               Q               K target score
       1      unknown [Kevin.hyatt - Mail_2].txt     Kevin.hyatt   TRUE 0.183
-      2  unknown [Kimberly.watson - Mail_3].txt     Kevin.hyatt  FALSE 0.209
-      3   unknown [Larry.campbell - Mail_1].txt     Kevin.hyatt  FALSE 0.234
-      4     unknown [Lindy.donoho - Mail_4].txt     Kevin.hyatt  FALSE 0.201
-      5       unknown [Liz.taylor - Mail_2].txt     Kevin.hyatt  FALSE 0.250
-      6      unknown [Kevin.hyatt - Mail_2].txt Kimberly.watson  FALSE 0.122
+      2  unknown [Kimberly.watson - Mail_3].txt     Kevin.hyatt  FALSE 0.175
+      3   unknown [Larry.campbell - Mail_1].txt     Kevin.hyatt  FALSE 0.195
+      4     unknown [Lindy.donoho - Mail_4].txt     Kevin.hyatt  FALSE 0.163
+      5       unknown [Liz.taylor - Mail_2].txt     Kevin.hyatt  FALSE 0.204
+      6      unknown [Kevin.hyatt - Mail_2].txt Kimberly.watson  FALSE 0.067
       7  unknown [Kimberly.watson - Mail_3].txt Kimberly.watson   TRUE 0.095
-      8   unknown [Larry.campbell - Mail_1].txt Kimberly.watson  FALSE 0.138
-      9     unknown [Lindy.donoho - Mail_4].txt Kimberly.watson  FALSE 0.146
-      10      unknown [Liz.taylor - Mail_2].txt Kimberly.watson  FALSE 0.148
-                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    unique_overlap
[…]
