+++ dev/pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   idiolect: idiolect.html
-last_built: 2025-03-14T14:43Z
+last_built: 2025-04-11T22:04Z
 urls:
   reference: https://andreanini.github.io/idiolect/reference
   article: https://andreanini.github.io/idiolect/articles
