### 2025-05-13 85486d9
add option to visualise heatmap as it is rather than just with sentences in lambdaG order
