+++ man/lambdaG_visualize.Rd
@@ -14,6 +14,7 @@ lambdaG_visualize(
   print = "",
   scale = "absolute",
   negative = FALSE,
+  order.by = "importance",
   cores = NULL
 )
 }
@@ -35,6 +36,8 @@ lambdaG_visualize(
 \item{scale}{A string indicating what scale to use to colour-code the text file. If "absolute" (default) then the raw \eqn{\lambda_G} is used; if "relative", then the z-score of \eqn{\lambda_G} over the Q data is used instead, thus showing relative importance.}
 
 \item{negative}{Logical. If TRUE then negative values of \eqn{\lambda_G} are color-coded in blue, otherwise (default) only the positive values of \eqn{\lambda_G} are displayed in red. This only applies to HTML output.}
+
+\item{order.by}{A string indicating the order of the output. If "importance" (default) then the output is ordered by sentence \eqn{\lambda_G} in descending order, otherwise the text is displayed and ordered as it appears.}
 
 \item{cores}{The number of cores to use for parallel processing (the default is one).}
 }
