+++ dev/pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 3.1.11
-pkgdown: 2.1.1
+pkgdown: 2.1.2
 pkgdown_sha: ~
 articles:
   idiolect: idiolect.html
-last_built: 2025-04-11T22:04Z
+last_built: 2025-05-13T16:29Z
 urls:
   reference: https://andreanini.github.io/idiolect/reference
   article: https://andreanini.github.io/idiolect/articles
