### 2025-09-19 #7575 @samhforbes
Hi @andreanini just checking in to see how it's all going
