### 2026-01-16 #7575 @andreanini
> It's probably outside the scope of this review, but I found the vignette in the documentation a bit tricky to follow in terms of mapping it to my own analysis case. But it might be less so for the primary users of the package if the nomenclature and language are common for the field. I'm coming at it from someone who understands the analysis methods and approach, but I am not in linguistics or a field that uses this analysis regularly (like English).
>
> There are three main parts where I think it could be clearer: defining the K, Q, and R sets (partly just formatting so definitions aren't in the middle of a paragraph, and partly showing an example more clearly), what the goal of the validation step is, and the language around computing and interpreting the likelihood statistic. I added issues related to these. The package works - this is just wanting to help it be used easily by it's intended audience.
>
> The only issue that needs to be resolved related to the review checklist is [#11](https://github.com/openjournals/joss-reviews/issues/11) above about community guidelines.

I have now improved the definitions of K/Q/R and provided examples, explained the validation in more detail, and fixed the wording around the LLRs.
