### 2026-01-23 #7575 @stefanocoretta
> @cmaimone and @stefanocoretta are you happy with the changes made? Apologies we're coming back to this after a delay so do let me know if this causes problems for you.

Yes, happy with the changes, no further comments!
