### 2026-02-28 #7575 @andreanini
Hi @samhforbes, all tasks assigned to me have now been completed. The new release is version 1.2.0 (already on CRAN) and this is the Zenodo archive: https://doi.org/10.5281/zenodo.18818044.

There are some references in the paper that are not correct or displayed properly. The information missing should be in the bib file but for some reason(perhaps the citation style) this does not appear in the paper. I have not changed anything yet as I can see that this is part of the next checklist.
