### 2026-03-01 #7575 @samhforbes
@andreanini It looks like the doi field needs populating for some of these entries
