### 2026-03-02 #7575 @andreanini
@samhforbes this is also done now, except for two papers for which I could not find the DOIs.
