### 2026-03-13 #7575 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
